Move new endpoints to batch service
